Code for calculating pore-slit overlaps in nm^2
Added code for calculating pore-slit overlaps. It creates a figure and an Excel file. In the Excel file, the results are in squared nanometers (nm^2). To run, overlap.py needs 2 Excel files: pores_corner_coords.xlsx and slits_corner_coords.xlsx

It takes less than 1.5 minutes to calculate overlapping areas for 601 angle values between 16 slits and 576 pores (however, the 576 pores are split into 16 groups, each group corresponds to a slit, so it's not like 1 slit is checked against all 576 pores for each angle value). 

Center of rotation is assumed to be in the center of the wafer.

Sometimes 2 pores are opened simultaneously by one slit. Some pores in a group are never opened by their corresponding slits.